OIDC: Move RP-initiated logout OP code into separate class
