+++ dev/com.ibm.ws.security.openidconnect.server/src/com/ibm/ws/security/openidconnect/web/OidcEndpointServices.java
@@ -4,7 +4,7 @@
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
- * 
+ *
  * SPDX-License-Identifier: EPL-2.0
  *
  * Contributors:
@@ -14,7 +14,6 @@ package com.ibm.ws.security.openidconnect.web;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.security.Principal;
 import java.util.Arrays;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -54,7 +53,6 @@ import com.ibm.websphere.ras.annotation.Sensitive;
 import com.ibm.websphere.ras.annotation.Trivial;
 import com.ibm.websphere.security.oauth20.AuthnContext;
 import com.ibm.ws.ffdc.annotation.FFDCIgnore;
-import com.ibm.ws.kernel.productinfo.ProductInfo;
 import com.ibm.ws.security.common.claims.UserClaims;
 import com.ibm.ws.security.oauth20.ProvidersService;
 import com.ibm.ws.security.oauth20.api.Constants;
@@ -63,16 +61,13 @@ import com.ibm.ws.security.oauth20.api.OidcOAuth20Client;
 import com.ibm.ws.security.oauth20.api.OidcOAuth20ClientProvider;
 import com.ibm.ws.security.oauth20.internal.AuthnContextImpl;
 import com.ibm.ws.security.oauth20.plugins.BaseClient;
-import com.ibm.ws.security.oauth20.plugins.OidcBaseClient;
 import com.ibm.ws.security.oauth20.plugins.jose4j.OidcUserClaims;
-import com.ibm.ws.security.oauth20.util.CacheUtil;
 import com.ibm.ws.security.oauth20.util.ConfigUtils;
 import com.ibm.ws.security.oauth20.util.OIDCConstants;
 import com.ibm.ws.security.oauth20.util.OidcOAuth20Util;
 import com.ibm.ws.security.oauth20.web.EndpointUtils;
 import com.ibm.ws.security.oauth20.web.OAuth20EndpointServices;
 import com.ibm.ws.security.oauth20.web.OAuth20Request.EndpointType;
-import com.ibm.ws.security.oauth20.web.OAuthClientTracker;
 import com.ibm.ws.security.oauth20.web.WebUtils;
 import com.ibm.ws.security.openidconnect.server.internal.HashUtils;
 import com.ibm.ws.security.openidconnect.server.internal.HttpUtils;
@@ -85,8 +80,6 @@ import com.ibm.wsspi.kernel.service.utils.AtomicServiceReference;
 import com.ibm.wsspi.kernel.service.utils.ConcurrentServiceReferenceMap;
 import com.ibm.wsspi.security.openidconnect.IDTokenMediator;
 import com.ibm.wsspi.security.openidconnect.UserinfoProvider;
-
-import io.openliberty.security.openidconnect.backchannellogout.BackchannelLogoutRequestHelper;
 
 @Component(service = { OidcEndpointServices.class }, name = "com.ibm.ws.security.openidconnect.web.OidcEndpointServices", immediate = true, configurationPolicy = ConfigurationPolicy.IGNORE, property = "service.vendor=IBM")
 public class OidcEndpointServices extends OAuth20EndpointServices {
@@ -369,190 +362,11 @@ public class OidcEndpointServices extends OAuth20EndpointServices {
      * - delete LTPAToken cookie.
      * - delete refresh token from tokencache if id_token_hint is present.
      * - redirect a request to a URL which is specified by post_logout_redirect_uri
-     *
-     * @param oauth20provider  extracted from the request
-     * @param oidcServerConfig is the object of oidc server configuration object
-     * @param request          is the incoming HttpServletRequest
-     * @param response         WAS OIDC response for a given provider
-     *
-     * @throws IOException
      */
-    @FFDCIgnore(IDTokenValidationFailedException.class)
     protected void processEndSession(OAuth20Provider oauth20provider, OidcServerConfig oidcServerConfig, HttpServletRequest request,
                                      HttpServletResponse response) throws ServletException, IOException {
-        Principal user = request.getUserPrincipal();
-        String idTokenString = request.getParameter(OIDCConstants.OIDC_LOGOUT_ID_TOKEN_HINT);
-        String redirectUri = request.getParameter(OIDCConstants.OIDC_LOGOUT_REDIRECT_URI);
-        String clientId = request.getParameter(OIDCConstants.OIDC_LOGOUT_CLIENT_ID);
-        OAuth20Token cachedIdToken = null;
-        if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-            Tr.debug(tc, "id_token_hint : " + idTokenString + " post_logout_redirect_uri : " + redirectUri + " client_id : " + clientId);
-        }
-        if (idTokenString != null && idTokenString.length() == 0) {
-            idTokenString = null;
-        }
-        boolean continueLogoff = true;
-
-        // lookup idtoken cache first.
-        OAuth20TokenCache tokenCache = null;
-        if (idTokenString != null) {
-            tokenCache = oauth20provider.getTokenCache();
-            if (tokenCache != null) {
-                String hash = HashUtils.digest(idTokenString);
-                if (hash != null) {
-                    cachedIdToken = tokenCache.get(hash);
-                    // if idToken is found, this is valid.
-                    if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                        Tr.debug(tc, "idToken : " + cachedIdToken);
-                    }
-                } else {
-                    Tr.error(tc, "OIDC_SERVER_IDTOKEN_VERIFY_ERR", new Object[] { "IDTokenValidatonFailedException" });
-                    continueLogoff = false;
-                }
-            }
-        }
-
-        String userName = ((user == null) ? null : user.getName());
-        String tokenUsername = ((cachedIdToken == null) ? null : cachedIdToken.getUsername());
-        clientId = ((cachedIdToken == null) ? clientId : cachedIdToken.getClientId());
-
-        if (idTokenString != null && cachedIdToken == null && continueLogoff) {
-            // if it's not there parse the idTokenString and validate signature.
-            JWT jwt = null;
-            try {
-                jwt = createJwt(idTokenString, oauth20provider, oidcServerConfig);
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                    Tr.debug(tc, "JWT : " + jwt);
-                }
-                //if (jwt.verify()) {
-                if (jwt.verifySignatureOnly()) {
-                    tokenUsername = JsonTokenUtil.getSub(jwt.getPayload());
-                    clientId = JsonTokenUtil.getAud(jwt.getPayload());
-                } else {
-                    Tr.error(tc, "OIDC_SERVER_IDTOKEN_VERIFY_ERR", new Object[] { "IDTokenValidatonFailedException" });
-                    continueLogoff = false;
-                }
-            } catch (IDTokenValidationFailedException ivfe) {
-                Throwable cause = ivfe.getCause();
-                if (cause != null && cause instanceof IllegalStateException) {
-                    // this error can be ignored, since this is due to exp, iat expiration.
-                    // extract sub.
-                    try {
-                        JWTPayload payload = JsonTokenUtil.getPayload(idTokenString);
-                        if (payload != null) {
-                            tokenUsername = JsonTokenUtil.getSub(payload);
-                            clientId = JsonTokenUtil.getAud(payload);
-                        }
-                    } catch (Exception e) {
-                        Tr.error(tc, "OIDC_SERVER_IDTOKEN_VERIFY_ERR", new Object[] { e });
-                        continueLogoff = false;
-                    }
-                } else {
-                    Tr.error(tc, "OIDC_SERVER_IDTOKEN_VERIFY_ERR", new Object[] { ivfe });
-                    continueLogoff = false;
-                }
-            } catch (Exception e) {
-                Tr.error(tc, "OIDC_SERVER_IDTOKEN_VERIFY_ERR", new Object[] { e });
-                continueLogoff = false;
-            }
-        }
-
-        if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-            Tr.debug(tc, "login username : " + userName + " IDToken username : " + tokenUsername);
-        }
-
-        if (userName != null && tokenUsername != null && !userName.equals(tokenUsername)) {
-            // user mismatch, abort
-            Tr.error(tc, "OIDC_SERVER_USERNAME_MISMATCH_ERR", new Object[] { userName, tokenUsername });
-            continueLogoff = false;
-        }
-
-        if (continueLogoff) {
-            if (cachedIdToken != null && tokenCache != null) {
-                // delete refreshtoken.
-                CacheUtil cu = new CacheUtil(tokenCache);
-                OAuth20Token refreshToken = cu.getRefreshToken(cachedIdToken);
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                    Tr.debug(tc, "refreshToken : " + refreshToken);
-                }
-                if (refreshToken != null) {
-                    tokenCache.remove(refreshToken.getTokenString());
-                }
-            }
-            //@AV999-092821
-//            if (user != null) {
-//                // logout deletes ltpatoken cookie and oidc_bsc cookie.
-//                request.logout();
-//            }
-        }
-
-        if (!continueLogoff) {
-            // this is an error condition. display an error page.
-            redirectUri = request.getContextPath() + "/end_session_error.html";
-        } else {
-            if (redirectUri == null) {
-                // no redirectUri is set, use default.
-                redirectUri = request.getContextPath() + "/end_session_logout.html";
-            } else {
-                try {
-                    String[] uris = getPostLogoutRedirectUris(oauth20provider, clientId);
-                    if (!containUri(redirectUri, uris)) {
-                        // post_logout_redirect_uri is not a member of post_logout_redirect_uris, force to redirect to the default logout page.
-                        if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                            if (clientId == null) {
-                                Tr.debug(tc,
-                                         "postLogoutRedirectUri value cannot be identified because client id is not set. Most likely this is because the id_token_hint parameter is not set or invalid.");
-                            }
-                        }
-                        Tr.error(tc, "OIDC_SERVER_LOGOUT_REDIRECT_URI_MISMATCH", new Object[] { redirectUri, printArray(uris), clientId });
-                        redirectUri = request.getContextPath() + "/end_session_logout.html";
-
-                    }
-                } catch (OidcServerException ose) {
-                    // this should not happen.
-                    Tr.error(tc, "OIDC_SERVER_IDTOKEN_VERIFY_ERR", new Object[] { ose });
-                    // this is an error condition. display an error page.
-                    redirectUri = request.getContextPath() + "/end_session_error.html";
-                }
-            }
-        }
-        if (oauth20provider.isTrackOAuthClients()) {
-            redirectUri = updateRedirectUriWithTrackedOAuthClients(request, response, oauth20provider, redirectUri);
-        }
-        //@AV999-092821
-        request.setAttribute("OIDC_END_SESSION_REDIRECT", redirectUri);
-        if (continueLogoff) {
-            if (user != null) {
-                // logout deletes ltpatoken cookie and oidc_bsc cookie.
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()) {
-                    Tr.debug(tc, "save  OIDC_END_SESSION_REDIRECT uri in op end_session : " + redirectUri);
-                }
-                //if during the servlet request logout, if the other logouts are in play, then we may not want to redirect here in that case
-                request.logout();
-            } else {
-                // request.logout() will send back-channel logout requests via the LogoutService OSGi service. Since request.logout()
-                // is only called in the above block if user != null, we need to make sure back-channel logout requests are still sent
-                // based on the id_token_hint if a user Principal isn't available
-                sendBackchannelLogoutRequests(request, oidcServerConfig, userName, idTokenString);
-            }
-        }
-        if (request.getAttribute("OIDC_END_SESSION_REDIRECT") != null) {
-            request.removeAttribute("OIDC_END_SESSION_REDIRECT");
-            if (tc.isDebugEnabled()) {
-                Tr.debug(tc, "OIDC _SSO OP redirecting to [" + redirectUri + "]");
-            }
-            response.sendRedirect(redirectUri);
-        }
-    }
-
-    String updateRedirectUriWithTrackedOAuthClients(HttpServletRequest request, HttpServletResponse response, OAuth20Provider provider, String redirectUri) {
-        OAuthClientTracker clientTracker = new OAuthClientTracker(request, response, provider);
-        return clientTracker.updateLogoutUrlAndDeleteCookie(redirectUri);
-    }
-
-    void sendBackchannelLogoutRequests(HttpServletRequest request, OidcServerConfig oidcServerConfig, String userName, String idTokenString) {
-        BackchannelLogoutRequestHelper bclRequestCreator = new BackchannelLogoutRequestHelper(request, oidcServerConfig);
-        bclRequestCreator.sendBackchannelLogoutRequests(userName, idTokenString);
+        OidcRpInitiatedLogout rpInitiatedLogout = new OidcRpInitiatedLogout(this, oauth20provider, oidcServerConfig, request, response);
+        rpInitiatedLogout.processEndSession();
     }
 
     /**
@@ -583,7 +397,7 @@ public class OidcEndpointServices extends OAuth20EndpointServices {
      * construct JWT which is a super class of IDToken from IdTokenString.
      * JWT class is used in order to just perform signature validation.
      *
-     * @param oauth20provider  extracted from the request
+     * @param oauth20provider extracted from the request
      * @param oidcServerConfig is the object of oidc server configurations
      * @throws OidcServerException
      *
@@ -626,47 +440,6 @@ public class OidcEndpointServices extends OAuth20EndpointServices {
             sharedKey = baseClient.getClientSecret();
         }
         return sharedKey;
-    }
-
-    /**
-     * get PostLogoutRedirectUris
-     *
-     * @param oauth20provider extracted from the request
-     * @param clientId
-     * @throws OidcServerException
-     *
-     */
-    String[] getPostLogoutRedirectUris(OAuth20Provider oauth20provider, String clientId) throws OidcServerException {
-        String[] uris = null;
-        if (clientId != null) {
-            OidcOAuth20ClientProvider clientProvider = oauth20provider.getClientProvider();
-            OidcOAuth20Client oauth20Client = clientProvider.get(clientId);
-            if (oauth20Client instanceof OidcBaseClient) {
-                OidcBaseClient baseClient = (OidcBaseClient) oauth20Client;
-                uris = OidcOAuth20Util.getStringArray(baseClient.getPostLogoutRedirectUris());
-            }
-        }
-        return uris;
-    }
-
-    /**
-     * get check whether the given string contains in the given JsonArray.
-     *
-     * @param uri  String.
-     * @param uris String[]
-     *
-     */
-    boolean containUri(String uri, String[] uris) {
-        boolean contain = false;
-        if (uris != null && uris.length > 0 && uri != null) {
-            for (int i = 0; i < uris.length; i++) {
-                if (uri.equals(uris[i])) {
-                    contain = true;
-                    break;
-                }
-            }
-        }
-        return contain;
     }
 
     /**
@@ -928,10 +701,10 @@ public class OidcEndpointServices extends OAuth20EndpointServices {
      * Return the JSONObject that will be returned for the userinfo endpoint, this method invokes the userinfo provider SPI
      * that has being installed on liberty
      *
-     * @param accessToken      the OAuth20Token used to get authentication context
+     * @param accessToken the OAuth20Token used to get authentication context
      * @param userinfoProvider the implementation of userinfoProvider which is installed at runtime
-     * @param request          the HTTPRequest for the userinfo endpoint
-     * @param response         the response for the userinfo endpoint request.
+     * @param request the HTTPRequest for the userinfo endpoint
+     * @param response the response for the userinfo endpoint request.
      * @return The JsonObject for userinfo endpoint
      * @throws IOException
      */
@@ -965,12 +738,12 @@ public class OidcEndpointServices extends OAuth20EndpointServices {
     /**
      * Get the JSONObject that will be returned for userinfo endpoint from the user registry
      *
-     * @param oauth20provider  The OAuth20Provider
+     * @param oauth20provider The OAuth20Provider
      * @param oidcServerConfig The OidcServerConfig
-     * @param request          The HttpServletRequest
-     * @param response         The HttpServletResponse
-     * @param accessToken      The OAuth20Token
-     * @param claims           The claims for this granted access
+     * @param request The HttpServletRequest
+     * @param response The HttpServletResponse
+     * @param accessToken The OAuth20Token
+     * @param claims The claims for this granted access
      * @param response
      * @throws IOException
      *
@@ -1026,31 +799,6 @@ public class OidcEndpointServices extends OAuth20EndpointServices {
         }
         response.setHeader(WWW_AUTHENTICATE_HEADER, header);
         response.setStatus(status);
-    }
-
-    /**
-     * Convert the String array to a String.
-     *
-     * @param value
-     * @return
-     */
-    @Trivial
-    private String printArray(String[] value) {
-        String result = null;
-        if (value != null && value.length > 0) {
-            StringBuffer buf = null;
-            for (int i = 0; i < value.length; i++) {
-                if (buf == null) {
-                    buf = new StringBuffer("[ ");
-                } else {
-                    buf.append(", ");
-                }
-                buf.append(value[i]);
-            }
-            buf.append(" ]");
-            result = buf.toString();
-        }
-        return result;
     }
 
     @FFDCIgnore({ IDTokenValidationFailedException.class, IllegalStateException.class })
